// packages/igdb-kit/src/generated/fields.ts
export const fieldSchemas: Record<EndpointName, Record<string, Omit<FieldSchema, "name">>> = {
 "age_rating_categories": {
  "id": {
   "type": "integer"
  },
// ...
 },
// ...
};